Load Chapters for Course
Loads the chapters for the course as well as passes the course title and
id to the Course Details Screen. This is the first step to loading
course Modules. Loading the user account information is critical to the
rest of the modules and removing hardcoded user information. This commit
also fixed the oauth issues to update the auth token correctly.